Moves Room->HipChatRoom, consistency on HipChat
